refactor(quiz): split matching and persona logic into modular engine with rules and helpers
