+++ backend/src/services/quiz/matching/rules/weighting.rules.ts
@@ -1,3 +1,9 @@
+// Extract category prefix from tagKey (e.g. "style:adventurous" → "style")
+export function getCategoryPrefix(tagKey: string): string {
+  const [prefix] = tagKey.split(':');
+  return prefix ?? 'default';
+}
+
 // ================================
 // POSITIVE WEIGHTING RULES
 // ================================
@@ -25,3 +31,47 @@ export const NEGATIVE_WEIGHTS = {
   medium: 1.2, // moderate penalty
   soft: 0.6, // small penalty
 };
+
+// ================================
+// CATEGORY WEIGHTING FOR SCORING (B + D)
+// ================================
+
+/**
+ * Category-level weights:
+ * - persona: strongest signal
+ * - style: strong
+ * - interest/activity: medium-strong
+ * - duration/budget/pace/region: medium
+ * - everything else: neutral
+ *
+ * High-importance tags (importance >= 4) are boosted slightly.
+ */
+export function getWeightedImportance(tagKey: string, importance: number): number {
+  const prefix = getCategoryPrefix(tagKey);
+
+  let baseFactor = 1.0;
+  switch (prefix) {
+    case 'persona':
+      baseFactor = 1.6;
+      break;
+    case 'style':
+      baseFactor = 1.4;
+      break;
+    case 'interest':
+    case 'activity':
+      baseFactor = 1.3;
+      break;
+    case 'duration':
+    case 'budget':
+    case 'pace':
+    case 'region':
+      baseFactor = 1.15;
+      break;
+    default:
+      baseFactor = 1.0;
+  }
+
+  const importanceBoost = importance >= 4 ? 1.15 : 1.0; // extra weight for "important" tags
+
+  return importance * baseFactor * importanceBoost;
+}
